services/asr-server/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
numpy==2.3.1
httpx==0.28.1
google-auth==2.40.3
transformers==5.4.0
torch==2.8.0
